Factor out api reference
